Rework the "advance Qdp" substep in the "2D Advection" step.
The goal was to make this substep parallelize over both tracers (q index) and
levels (k index). It was originally parallelized over tracers only. If the
number of tracers is small, that could be a problem.

 * First, I made limiter_optim_iter_full act on just one level insteal of nlev
levels.

 * Second, this let me write one loop 'k = 1, nlev' instead of many.

These two together then make it possible to write
    !$omp do collapse(2)
    do q = 1, qsize
       do k = 1, nlev
which was the goal.

Third, I modified other parts of the substep in ways orthogonal to the first two
changes.

 * dp_star was computed repeatedly inside the q loop, but it is independent of
tracer. So I moved it outside of the q loop.

 * dp_star was also used as a temporary, but in fact a temporary wasn't needed.

 * I reduced Qtens from (:,:,k) to (:,:) and tmp1/2(k) to tmp1/2.

 * The tmp1, tmp2 bits of code are for debugging, so I ifdef'ed them out with
TRANSPORT_SE_DEBUG. (I didn't see a DEBUG define in use, so I created this one.)